feat(queue): optimize per-event capacity worker and fix FIFO violations
