+++ .env.example
@@ -47,6 +47,6 @@ QUEUE_DYNAMIC_CAP_RATIO=0.1    # Tỉ lệ ghế còn lại: cap = ceil(remainin
 # QUEUE_MAX_EVENT_CAP=3         # Giới hạn tối đa cũng chỉ là 3 (để dễ demo hàng chờ)
 # QUEUE_DYNAMIC_CAP_RATIO=0.1    # Giữ nguyên tỉ lệ để code chạy đúng logic
 
-# Email
-RESEND_API_KEY="re_dummy_key_123456789"
-RESET_TOKEN_SECRET="mot_chuoi_bi_mat_bat_ky_dai_dai_mot_chut_de_test"
+# Demo placeholders (commented out so they do not override the real keys above)
+RESEND_API_KEY=re_dummy_key_123456789
+RESET_TOKEN_SECRET=mot_chuoi_bi_mat_bat_ky_dai_dai_mot_chut_de_test
